+++ src/main/resources/com/jiro4989/tkfm/Main.fxml
@@ -34,7 +34,36 @@
     </BorderPane>
   </center>
   <left>
-    <TitledPane fx:id="fileListPane" animated="false" collapsible="false" content="$null" expanded="true" prefHeight="2000.0" prefWidth="397.0" text="ファイルリスト">
+    <TitledPane animated="false" collapsible="false" content="$null" expanded="true" prefHeight="2000.0" prefWidth="397.0" text="ファイルリスト">
+      <HBox prefHeight="699.0" prefWidth="407.0" spacing="5.0" xmlns:fx="http://javafx.com/fxml/1" xmlns="http://javafx.com/javafx/2.2">
+        <children>
+          <ListView fx:id="fileListView" prefHeight="644.0" prefWidth="253.0" />
+          <AnchorPane id="AnchorPane" prefHeight="801.0" prefWidth="144.0">
+            <children>
+              <GridPane prefHeight="411.0" prefWidth="99.0" vgap="5.0" AnchorPane.leftAnchor="0.0" AnchorPane.rightAnchor="0.0" AnchorPane.topAnchor="0.0">
+                <children>
+                  <Button fx:id="insertButton" mnemonicParsing="false" prefHeight="99.0" prefWidth="335.0" text="一括挿入" GridPane.columnIndex="0" GridPane.rowIndex="0" />
+                  <Button fx:id="clearButton" mnemonicParsing="false" prefHeight="99.0" prefWidth="321.0" text="クリア" GridPane.columnIndex="0" GridPane.rowIndex="1" />
+                  <Button fx:id="listDeleteButton" mnemonicParsing="false" prefHeight="99.0" prefWidth="229.0" text="リスト削除" GridPane.columnIndex="0" GridPane.rowIndex="2" />
+                  <Button fx:id="listClearButton" mnemonicParsing="false" prefHeight="99.0" prefWidth="279.0" text="リストクリア" GridPane.columnIndex="0" GridPane.rowIndex="3" />
+                </children>
+                <columnConstraints>
+                  <ColumnConstraints hgrow="SOMETIMES" minWidth="10.0" prefWidth="100.0" />
+                </columnConstraints>
+                <rowConstraints>
+                  <RowConstraints minHeight="10.0" prefHeight="30.0" vgrow="SOMETIMES" />
+                  <RowConstraints minHeight="10.0" prefHeight="30.0" vgrow="SOMETIMES" />
+                  <RowConstraints minHeight="10.0" prefHeight="30.0" vgrow="SOMETIMES" />
+                  <RowConstraints minHeight="10.0" prefHeight="30.0" vgrow="SOMETIMES" />
+                </rowConstraints>
+              </GridPane>
+            </children>
+          </AnchorPane>
+        </children>
+        <padding>
+          <Insets bottom="5.0" left="5.0" right="5.0" top="5.0" />
+        </padding>
+      </HBox>
       <BorderPane.margin>
         <Insets bottom="5.0" left="5.0" right="5.0" top="5.0" fx:id="x1" />
       </BorderPane.margin>
